DB generation scripts and associated command have been updated.

+++ fetchNCBIBLASTDB.sh
@@ -1,16 +1,26 @@
+# Set number of processor cores used for computation
+export NCPU=`grep -c processor /proc/cpuinfo`
+# Set PREFIX
+if test -z $PREFIX; then
+PREFIX=/usr/local || exit $?
+fi
 # get nt database from NCBI
 mkdir -p blastdb || exit $?
 cd blastdb || exit $?
-wget -c -e robots=off -r -l1 -np https://ftp.ncbi.nih.gov/blast/db/ -A nt.??.tar.gz.md5 -R env_nt.??.tar.gz.md5 || exit $?
-wget -c -e robots=off -r -l1 -np https://ftp.ncbi.nih.gov/blast/db/ -A nt.??.tar.gz -R env_nt.??.tar.gz || exit $?
-wget -c https://ftp.ncbi.nih.gov/blast/db/taxdb.tar.gz.md5 || exit $?
-wget -c https://ftp.ncbi.nih.gov/blast/db/taxdb.tar.gz || exit $?
-find ftp.ncbi.nih.gov -name nt.??.tar.gz | xargs -I {} mv {} ./
-find ftp.ncbi.nih.gov -name nt.??.tar.gz.md5 | xargs -I {} mv {} ./
-rm -rf ftp.ncbi.nih.gov || exit $?
-ls *.md5 | xargs -L 1 -P 16 -I {} sh -c "md5sum -c {} || exit $?" || exit $?
+aria2c https://ftp.ncbi.nih.gov/blast/db/ -o index.html || exit $?
+grep -o -P 'nt_(euk|prok).\d+.tar.gz.md5' index.html | sort -u | perl -npe 's/^/https:\/\/ftp.ncbi.nih.gov\/blast\/db\//' > md5list.txt || exit $?
+aria2c -c -i md5list.txt -j 3 -x 1 || exit $?
+rm md5list.txt || exit $?
+grep -o -P 'nt_(euk|prok).\d+.tar.gz' index.html | sort -u | perl -npe 's/^/https:\/\/ftp.ncbi.nih.gov\/blast\/db\//' > targzlist.txt || exit $?
+aria2c -c -i targzlist.txt -j 3 -x 1 || exit $?
+rm targzlist.txt || exit $?
+rm index.html || exit $?
+aria2c -c https://ftp.ncbi.nih.gov/blast/db/taxdb.tar.gz.md5 || exit $?
+aria2c -c https://ftp.ncbi.nih.gov/blast/db/taxdb.tar.gz || exit $?
+ls *.md5 | xargs -P $NCPU -I {} sh -c "md5sum -c {} || exit $?" || exit $?
 rm *.md5 || exit $?
-ls *.tar.gz | xargs -L 1 -P 16 -I {} sh -c "tar -xzf {} || exit $?" || exit $?
-chmod 644 *.tar.gz || exit $?
-rm *.tar.gz || exit $?
+ls *.tar.gz | xargs -P $NCPU -I {} sh -c "tar -xzf {} || exit $?" || exit $?
+chmod 644 *.tar.gz || sudo chmod 644 *.tar.gz || exit $?
+rm *.tar.gz || sudo rm *.tar.gz || exit $?
+BLASTDB=./ $PREFIX/share/claident/bin/blastdb_aliastool -dbtype nucl -out nt -title nt -dblist 'nt_euk nt_prok' || exit $?
 cd .. || exit $?
